BLD : added SNAPSHOT flag
having the version tagged wit the git build makes reloading
annoying as the egg infrastructure adds the install path (which
includes the git SHA) to sys.path python will only find the old version
of modules on `reload` even if a new version is installed.

The ability to tack git versions onto the version might still be
useful for building trackable snapshots.